+++ yomikomi_prj/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod render;
 pub mod api;
 pub mod route;
+pub mod model;
